Промежуточная фиксация

+++ oneswiss-agent/Services/MaintenanceTasks/MaintenanceTaskExecutor.cs
@@ -1,10 +1,13 @@
 using System.CommandLine.Parsing;
+using System.Text.RegularExpressions;
 using OneSwiss.Agent.Extensions;
 using OneSwiss.Common.DTO;
 using OneSwiss.Common.DTO.MaintenanceTasks;
 using OneSwiss.Common.Extensions;
 using OneSwiss.Common.Models.MaintenanceTasks;
 using OneSwiss.OneScript;
+using OneSwiss.V8.Designer.Agent;
+using OneSwiss.V8.Designer.Batch;
 using OneSwiss.V8.Platform.RemoteAdministration;
 
 namespace OneSwiss.Agent.Services.MaintenanceTasks;
@@ -57,12 +60,14 @@ public class MaintenanceTaskExecutor : BackgroundService
     private async Task StartMaintenanceTask(MaintenanceTaskDto task, CancellationToken cancellationToken)
     {
         var v8Files = await SaveTaskV8Files(task, cancellationToken);
+
+        /*var hasStepsForDesigner = HasStepsForDesigner(task);
+        var canUseDesignerAgent = CanUseDesignerAgent(task);*/
+        var hasStepsForDesigner = true;
+        var canUseDesignerAgent = false;
         
         await Parallel.ForEachAsync(task.InfoBases, cancellationToken, async (infoBase, stoppingToken) =>
         {
-            // Перекопируем файлы для каждого задания, т.к. 1С на кой-то хрен нужен монопольный доступ к файлам CF, CFE, CFU
-            var localFiles = CopyFilesForInfoBase(v8Files);
-            
             var log = new List<MaintenanceStepLogItemDto>();
             
             var context = new MaintenanceStepContext
@@ -70,9 +75,11 @@ public class MaintenanceTaskExecutor : BackgroundService
                 Task = task,
                 InfoBase = infoBase,
                 Log = log,
-                Files = localFiles,
-                Step = task.Steps.GetRootStep()
+                Step = task.Steps.GetRootStep(),
+                UseDesignerAgent = canUseDesignerAgent
             };
+
+            OnecV8BatchMode? agent = null;
 
             try
             {
@@ -84,6 +91,29 @@ public class MaintenanceTaskExecutor : BackgroundService
 
                 if (!context.Platform.HasOnecV8)
                     throw new Exception("Для платформы агента не установлен конфигуратор");
+                
+                var localFilesFolder = Path.Join(Path.GetTempPath(), Guid.NewGuid().ToString(), "0");
+                Directory.CreateDirectory(localFilesFolder);
+                
+                if (hasStepsForDesigner && canUseDesignerAgent)
+                {
+                    agent = context.StartDesignerAgent(Path.GetDirectoryName(localFilesFolder)!);
+                
+                    context.DesignerAgentClient = new DesignerAgentClient(
+                        context.InfoBase.Credentials?.User ?? "",
+                        context.InfoBase.Credentials?.Password ?? "");
+
+                    if (!await context.DesignerAgentClient.WaitAgentAvailable(TimeSpan.FromMinutes(5)))
+                        throw new Exception("Таймаут подключения к агенту конфигуратора");
+
+                    await context.DesignerAgentClient.Connect(stoppingToken);
+                    await context.DesignerAgentClient.ConnectIb();
+                }
+                
+                // Перекопируем файлы для каждого задания, т.к. 1С на кой-то хрен нужен монопольный доступ к файлам CF, CFE, CFU
+                // и сделаем это обязательно после запуска агента конфигуратора, иначе он затрет их при запуске
+                var localFiles = CopyFilesForInfoBase(v8Files, localFilesFolder);
+                context.Files = localFiles;
 
                 while (!stoppingToken.IsCancellationRequested)
                 {
@@ -97,7 +127,7 @@ public class MaintenanceTaskExecutor : BackgroundService
                     {
                         try
                         {
-                            HandleTaskStepNode(context);
+                            await HandleTaskStepNode(context);
 
                             if (context.Step.LeftStepId is not null)
                                 context.Step = task.Steps.GetStep(context.Step.LeftStepId);
@@ -116,7 +146,7 @@ public class MaintenanceTaskExecutor : BackgroundService
                     }
                     else
                     {
-                        HandleTaskStepNode(context);
+                        await HandleTaskStepNode(context);
 
                         if (context.Step.LeftStepId is not null)
                             context.Step = task.Steps.GetStep(context.Step.LeftStepId);
@@ -135,21 +165,36 @@ public class MaintenanceTaskExecutor : BackgroundService
             }
             finally
             {
-                DeleteV8Files(localFiles);
+                if (context.UseDesignerAgent)
+                {
+                    try
+                    {
+                        context.DesignerAgentClient?.DisconnectIb();
+                    }
+                    catch
+                    {
+                        // ignored
+                    }
+
+                    context.DesignerAgentClient?.Dispose();
+                }
+                    
+                agent?.Dispose();
+                DeleteV8Files(context.Files, true);
             }
         });
         
-        DeleteV8Files(v8Files);
+        DeleteV8Files(v8Files, false);
     }
 
-    private static Dictionary<Guid, string> CopyFilesForInfoBase(Dictionary<Guid, string> files)
+    private static Dictionary<Guid, string> CopyFilesForInfoBase(Dictionary<Guid, string> files, string folder)
     {
         var result = new Dictionary<Guid, string>();
 
         foreach (var file in files)
         {
             var extension = Path.GetExtension(file.Value);
-            var path = Path.Join(Path.GetTempPath(), $"{Guid.NewGuid()}{extension}");
+            var path = Path.Join(folder, $"{Guid.NewGuid()}{extension}");
             File.Copy(file.Value, path);
             
             result.Add(file.Key, path);
@@ -158,7 +203,7 @@ public class MaintenanceTaskExecutor : BackgroundService
         return result;
     }
 
-    private static void HandleTaskStepNode(MaintenanceStepContext context)
+    private static async Task HandleTaskStepNode(MaintenanceStepContext context)
     {
         AddLogItem(context, $"Обработка шага \"{context.Step.Kind.GetDisplay()}\"");
 
@@ -174,16 +219,16 @@ public class MaintenanceTaskExecutor : BackgroundService
                 UnlockConnections(context);
                 break;
             case MaintenanceStepKind.LoadExtension:
-                LoadExtension(context);
+                await LoadExtension(context);
                 break;
             case MaintenanceStepKind.UpdateConfiguration:
                 UpdateConfiguration(context);
                 break;
             case MaintenanceStepKind.LoadConfiguration:
-                LoadConfiguration(context);
+                await LoadConfiguration(context);
                 break;
             case MaintenanceStepKind.DeleteExtension:
-                DeleteExtension(context);
+                await DeleteExtension(context);
                 break;
             case MaintenanceStepKind.StartExternalDataProcessor:
                 StartExternalDataProcessor(context);
@@ -218,7 +263,7 @@ public class MaintenanceTaskExecutor : BackgroundService
         CancellationToken cancellationToken)
     {
         await using var scope = _serviceProvider.CreateAsyncScope();
-        using var downloader = scope.ServiceProvider.GetRequiredService<V8FilesDownloader>();
+        using var downloader = scope.ServiceProvider.GetRequiredService<FilesDownloader>();
 
         var filesToDownload = task.Steps
             .Where(c => c.File is not null)
@@ -229,18 +274,30 @@ public class MaintenanceTaskExecutor : BackgroundService
         return await downloader.Download(_serverConnection, filesToDownload, cancellationToken);
     }
 
-    private static void DeleteV8Files(Dictionary<Guid, string> files)
+    private static void DeleteV8Files(Dictionary<Guid, string> files, bool isLocal)
     {
-        foreach (var file in files.Values)
+        if (files.Count == 0)
+            return;
+        
+        // Если это локальные файлы, то можем затереть сразу весь каталог
+        if (isLocal)
         {
-            try
+            var folder = Path.GetDirectoryName(Path.GetDirectoryName(files.Values.First()));
+            Directory.Delete(folder!, true);
+        }
+        else
+        {
+            foreach (var file in files.Values)
             {
-                if (File.Exists(file))
-                    File.Delete(file);
-            }
-            catch
-            {
-                // ignore
+                try
+                {
+                    if (File.Exists(file))
+                        File.Delete(file);
+                }
+                catch
+                {
+                    // ignore
+                }
             }
         }
     }
@@ -299,20 +356,32 @@ public class MaintenanceTaskExecutor : BackgroundService
             context.InfoBase.Credentials?.Password ?? "");
     }
     
-    private static void LoadExtension(MaintenanceStepContext context)
+    private static async Task LoadExtension(MaintenanceStepContext context)
     {
         var filePath = context.Files[context.Step.File!.Id];
+
+        if (context.UseDesignerAgent)
+        {
+            await context.DesignerAgentClient!.LoadExtension(Path.GetFileName(filePath), context.Step.ExtensionName);
+            AddLogItem(context, $"Загрузка расширения \"{context.Step.ExtensionName}\" выполнена");
+            
+            context.DesignerAgentClient!.UpdateDbCfgExtension(context.Step.ExtensionName);
+            await context.DesignerAgentClient.ReadMessagesTillSuccess(message =>
+                LogDesignerAgentMessage(context, message));
+        }
+        else
+        {
+            using var batch = context.GetBatchDesigner();
+            batch.LoadExtension(
+                context.Step.ExtensionName, 
+                filePath, 
+                context.InfoBase.Credentials?.User ?? "", 
+                context.InfoBase.Credentials?.Password ?? "", 
+                context.AccessCode,
+                true);
         
-        using var batch = context.GetBatchDesigner();
-        batch.LoadExtension(
-            context.Step.File!.Name, 
-            filePath, 
-            context.InfoBase.Credentials?.User ?? "", 
-            context.InfoBase.Credentials?.Password ?? "", 
-            context.AccessCode,
-            true);
-        
-        AddLogItem(context, batch.OutFileContent);
+            AddLogItem(context, batch.OutFileContent);
+        }
     }
     
     private static void ExecuteOneScript(MaintenanceStepContext context)
@@ -340,32 +409,70 @@ public class MaintenanceTaskExecutor : BackgroundService
         Directory.Delete(scriptPath, true);
     }
     
-    private static void DeleteExtension(MaintenanceStepContext context)
+    private static async Task DeleteExtension(MaintenanceStepContext context)
     {
-        using var batch = context.GetBatchDesigner();
-        batch.DeleteExtension(
-            context.Step.ExtensionName, 
-            context.InfoBase.Credentials?.User ?? "", 
-            context.InfoBase.Credentials?.Password ?? "", 
-            context.AccessCode,
-            true);
+        if (context.UseDesignerAgent)
+        {
+            var allExtensions = await context.DesignerAgentClient!.GetAllExtensions();
+            var extensionsToDeleting = allExtensions.Where(c => Regex.IsMatch(c.Name, context.Step.ExtensionName)).ToList();
+
+            foreach (var extension in extensionsToDeleting)
+            {
+                await context.DesignerAgentClient!.DeleteExtension(extension.Name);
+                AddLogItem(context, $"Расширение \"{extension.Name}\" удалено");
+            }
+        }
+        else
+        {
+            using var batchGet = context.GetBatchDesigner();
+            var allExtensions = batchGet.GetExtensionsList(
+                context.InfoBase.Credentials?.User ?? "", 
+                context.InfoBase.Credentials?.Password ?? "", 
+                context.AccessCode,
+                true);
+            
+            var extensionsToDeleting = allExtensions.Where(c => Regex.IsMatch(c, context.Step.ExtensionName)).ToList();
+            
+            foreach (var extension in extensionsToDeleting)
+            {
+                using var batchDeleting = context.GetBatchDesigner();
+                batchDeleting.DeleteExtension(
+                    extension, 
+                    context.InfoBase.Credentials?.User ?? "", 
+                    context.InfoBase.Credentials?.Password ?? "", 
+                    context.AccessCode,
+                    true);
         
-        AddLogItem(context, batch.OutFileContent);
+                AddLogItem(context, batchDeleting.OutFileContent);
+            }
+        }
     }
 
-    private static void LoadConfiguration(MaintenanceStepContext context)
+    private static async Task LoadConfiguration(MaintenanceStepContext context)
     {
         var filePath = context.Files[context.Step.File!.Id];
+
+        if (context.UseDesignerAgent)
+        {
+            await context.DesignerAgentClient!.LoadCfg(Path.GetFileName(filePath));
+            AddLogItem(context, "Загрузка конфигурации выполнена");
+            
+            context.DesignerAgentClient!.UpdateDbCfg();
+            await context.DesignerAgentClient.ReadMessagesTillSuccess(message =>
+                LogDesignerAgentMessage(context, message));
+        }
+        else
+        {
+            using var batch = context.GetBatchDesigner();
+            batch.LoadConfiguration(
+                filePath, 
+                context.InfoBase.Credentials?.User ?? "", 
+                context.InfoBase.Credentials?.Password ?? "", 
+                context.AccessCode,
+                true);
         
-        using var batch = context.GetBatchDesigner();
-        batch.LoadConfiguration(
-            filePath, 
-            context.InfoBase.Credentials?.User ?? "", 
-            context.InfoBase.Credentials?.Password ?? "", 
-            context.AccessCode,
-            true);
-        
-        AddLogItem(context, batch.OutFileContent);
+            AddLogItem(context, batch.OutFileContent);
+        }
     }
     
     private static void UpdateConfiguration(MaintenanceStepContext context)
@@ -395,6 +502,38 @@ public class MaintenanceTaskExecutor : BackgroundService
             context.AccessCode,
             true);
     }
+
+    private static bool HasStepsForDesigner(MaintenanceTaskDto task)
+    {
+        var has = false;
+
+        foreach (var step in task.Steps)
+        {
+            if (step.Kind is MaintenanceStepKind.DeleteExtension or 
+                MaintenanceStepKind.LoadConfiguration or 
+                MaintenanceStepKind.LoadExtension or 
+                MaintenanceStepKind.UpdateConfiguration)
+                has = true;
+        }
+        
+        return has;
+    }
+
+    private static void LogDesignerAgentMessage(MaintenanceStepContext context, DesignerAgentMessage message)
+    {
+        if (message.Type == "log")
+        {
+            if (message.Message.StartsWith("(!)"))
+                AddLogItem(context, message.Message[3..], true);
+            else
+                AddLogItem(context, message.Message);
+        }
+        else
+            throw new Exception($"Неожиданный тип сообщения: {message.Type}");
+    }
+
+    private static bool CanUseDesignerAgent(MaintenanceTaskDto task)
+        => task.Steps.FirstOrDefault(c => c.Kind == MaintenanceStepKind.UpdateConfiguration) == null;
 
     public override void Dispose()
     {
